+++ tests/files/s7.jl
@@ -1,0 +1,8 @@
+baremodule Bare
+
+mutable struct MPoint
+   x
+   y
+end
+
+end